+++ src/components/reflection-detail/reflection-setting/ReflectionSettingHeader.tsx
@@ -1,9 +1,7 @@
-import { useState } from "react";
-import Link from "next/link";
-import { Box, Button, Fade, Popper } from "@mui/material";
+import { Box } from "@mui/material";
 import { KebabButtonPopupContainer } from "../../ui/shared/popup";
+import HeaderTableOfContents from "../table-of-contents/HeaderTableOfContents";
 import ToEditPageButton from "./ToEditPageButton";
-import { theme } from "@/src/utils/theme";
 
 type ReflectionSettingHeaderProps = {
   username: string;
@@ -20,16 +18,6 @@ const ReflectionSettingHeader: React.FC<ReflectionSettingHeaderProps> = ({
   isCurrentUser,
   tocArray
 }) => {
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(anchorEl ? null : event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
   return (
     <>
       <Box
@@ -49,70 +37,7 @@ const ReflectionSettingHeader: React.FC<ReflectionSettingHeaderProps> = ({
           boxShadow={{ xs: "0px 0.7px 1px rgba(0, 0, 0, 0.1)", md: "none" }}
         >
           <Box mt={0.5}>
-            <Button
-              onClick={handleClick}
-              onBlur={handleClose}
-              sx={{
-                display: { xs: "block", md: "none" },
-                width: "90px",
-                border: "none",
-                alignItems: "center",
-                whiteSpace: "nowrap",
-                color: theme.palette.grey[600]
-              }}
-            >
-              目次
-            </Button>
-            <Popper
-              open={Boolean(anchorEl)}
-              anchorEl={anchorEl}
-              transition
-              sx={{ zIndex: 2 }}
-            >
-              {({ TransitionProps }) => (
-                <Fade {...TransitionProps} timeout={250}>
-                  <Box boxShadow={1} borderRadius={2} bgcolor={"white"}>
-                    {tocArray.map((item) => (
-                      <Box
-                        key={item.href}
-                        display={"flex"}
-                        flexDirection={"column"}
-                        px={2}
-                        py={0.75}
-                      >
-                        <Link
-                          href={item.href}
-                          onClick={(e) => {
-                            e.preventDefault();
-                            // MEMO: ヘッダーを含めた時にロジックまで飛ばすロジック
-                            const targetElement = document.querySelector(
-                              item.href
-                            );
-                            if (targetElement) {
-                              const headerOffset = 55;
-                              const elementPosition =
-                                targetElement.getBoundingClientRect().top;
-                              const offsetPosition =
-                                elementPosition + window.scrollY - headerOffset;
-                              window.scrollTo({
-                                top: offsetPosition,
-                                behavior: "smooth"
-                              });
-                            }
-                          }}
-                          style={{
-                            textDecoration: "none",
-                            color: theme.palette.grey[600]
-                          }}
-                        >
-                          {item.title}
-                        </Link>
-                      </Box>
-                    ))}
-                  </Box>
-                </Fade>
-              )}
-            </Popper>
+            <HeaderTableOfContents tocArray={tocArray} />
           </Box>
           <Box mt={0.5}>
             <KebabButtonPopupContainer
